+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 - Reconnect to file watcher when disconnected from WebSocket server ([#130](https://github.com/marp-team/marp-cli/pull/130))
 - Change port number for file watcher from 52000 to 37717 ([#135](https://github.com/marp-team/marp-cli/issues/135), [#137](https://github.com/marp-team/marp-cli/pull/137))
+- Upgrade dependent packages to the latest version ([#138](https://github.com/marp-team/marp-cli/pull/138))
 
 ## v0.12.1 - 2019-07-13